refactor: move dto package to application level - Item controller, service test code modify
