+++ src/main/java/com/delivery/justonebite/item/application/dto/response/ItemResponse.java
@@ -1,4 +1,4 @@
-package com.delivery.justonebite.item.presentation.dto.response;
+package com.delivery.justonebite.item.application.dto.response;
 
 import com.delivery.justonebite.item.domain.entity.Item;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -6,9 +6,9 @@ import lombok.Builder;
 
 import java.util.UUID;
 
-@Schema(description = "상품 상세정보 조회 응답 DTO")
+@Schema(description = "상품 정보 조회 응답 DTO")
 @Builder
-public record ItemDetailResponse(
+public record ItemResponse(
     @Schema(description = "상품 고유 ID", example = "예시: a1b2c3d4-e5f6-4a7b-8c9d-0e1f2a3b4c5d")
     UUID itemId,
 
@@ -19,18 +19,14 @@ public record ItemDetailResponse(
     int price,
 
     @Schema(description = "상품의 image")
-    String image,
-
-    @Schema(description = "상품의 설명")
-    String description
+    String image
 ) {
-    public static ItemDetailResponse from(Item item) {
-        return ItemDetailResponse.builder()
+    public static ItemResponse from(Item item) {
+        return ItemResponse.builder()
             .itemId(item.getItemId())
             .name(item.getName())
             .price(item.getPrice())
             .image(item.getImage())
-            .description(item.getDescription())
             .build();
     }
 }
