+++ src/main/java/com/delivery/justonebite/item/domain/entity/Item.java
@@ -1,12 +1,11 @@
 package com.delivery.justonebite.item.domain.entity;
 
 import com.delivery.justonebite.global.common.entity.BaseEntity;
-import com.delivery.justonebite.item.presentation.dto.request.ItemUpdateRequest;
+import com.delivery.justonebite.item.application.dto.request.ItemUpdateRequest;
 import com.delivery.justonebite.shop.domain.entity.Shop;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 import lombok.experimental.SuperBuilder;
 import org.hibernate.annotations.SQLRestriction;
