+++ src/test/java/com/delivery/justonebite/item/domain/repository/ItemRepositoryTest.java
@@ -2,7 +2,7 @@ package com.delivery.justonebite.item.domain.repository;
 
 import com.delivery.justonebite.item.domain.entity.Item;
 
-import com.delivery.justonebite.item.presentation.dto.request.ItemUpdateRequest;
+import com.delivery.justonebite.item.application.dto.request.ItemUpdateRequest;
 import com.delivery.justonebite.shop.domain.entity.Shop;
 import com.delivery.justonebite.shop.domain.repository.ShopRepository;
 import com.delivery.justonebite.user.domain.entity.User;
